auth/requirements.txt
astroid==2.15.3
autopep8==2.0.2
click==8.1.3
dill==0.3.6
Flask==2.2.3
Flask-MySQLdb==1.0.1
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mysqlclient==2.1.1
parso==0.8.3
platformdirs==3.2.0
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.17.2
python-dotenv==1.0.0
tomlkit==0.11.7
Werkzeug==2.2.3
wrapt==1.15.0
